refactor(restaurantCategories.js):restaurantsData 네이밍 변경
식당정보데이터를 서버에서 불러오게 바뀌어 restaurantsData의 역할이 카테고리와 이미지 제공으로 바뀜에 따라 네이밍 변경
